module5-task1: разделил код на модули и сделал main.js точкой входа
