+++ index.html
@@ -233,8 +233,8 @@
   </template>
 
   <!-- Подключение JavaScript-файла -->
-  <script src="js/functions.js"></script>
-  <script src="js/main.js"></script>
+
+  <script type="module" src="js/main.js"></script>
 
 </body>
 </html>
